worker/requirements.txt
redis==5.2.0
rq==1.16.2
python-dotenv==1.0.1
